kafka/app/requirements.txt
confluent-kafka==1.8.2
Faker==13.6.0
pyyaml-6.0
requests-2.27.1
avro-1.11.0
